+++ ufe/src/css/main.css
@@ -1,0 +1,7 @@
+.center {
+  margin: 0 auto;
+  display: flex;
+  flex: 1;
+  justify-content: "center";
+  align-items: "center";
+}
